backend/requirements.txt
fastapi
uvicorn[standard]
deepeval
openai
pydantic
pydantic-settings
python-dotenv
scikit-learn
numpy
python-multipart
sqlalchemy
psycopg2-binary
alembic
bcrypt>=4.0.0
python-jose[cryptography]
cryptography
tiktoken
